chore: Rename MyProfileHeaderMyCollectionAndSavedWorks
