+++ src/app/Scenes/MyCollection/MyCollection.tsx
@@ -42,7 +42,7 @@ import { Button, Flex, Separator, Spacer } from "palette"
 import React, { useContext, useEffect, useRef, useState } from "react"
 import { createPaginationContainer, graphql, QueryRenderer, RelayPaginationProp } from "react-relay"
 import { useTracking } from "react-tracking"
-import { Tab } from "../MyProfile/MyProfileHeaderMyCollectionAndSavedWorks"
+import { Tab } from "../MyProfile/MyProfile"
 import { ARTWORK_LIST_IMAGE_SIZE } from "./Components/MyCollectionArtworkListItem"
 import { MyCollectionArtworks } from "./MyCollectionArtworks"
 import { MyCollectionArtworkUploadMessages } from "./Screens/ArtworkForm/MyCollectionArtworkUploadMessages"
