+++ src/app/Scenes/MyCollection/MyCollectionArtworks.tsx
@@ -13,7 +13,7 @@ import { FlatList, LayoutAnimation, NativeScrollEvent, NativeSyntheticEvent } fr
 import { graphql, RelayPaginationProp } from "react-relay"
 import { useTracking } from "react-tracking"
 import { useScreenDimensions } from "shared/hooks"
-import { Tab } from "../MyProfile/MyProfileHeaderMyCollectionAndSavedWorks"
+import { Tab } from "../MyProfile/MyProfile"
 import { MyCollectionArtworkList } from "./Components/MyCollectionArtworkList"
 import { MyCollectionSearchBar } from "./Components/MyCollectionSearchBar"
 import { MyCollectionArtworkEdge } from "./MyCollection"
