+++ src/app/Scenes/MyCollection/Screens/Insights/ActivateMoreMarketInsightsBanner.tsx
@@ -1,6 +1,6 @@
 import { ActionType, OwnerType } from "@artsy/cohesion"
 import { navigate, popToRoot } from "app/navigation/navigate"
-import { Tab } from "app/Scenes/MyProfile/MyProfileHeaderMyCollectionAndSavedWorks"
+import { Tab } from "app/Scenes/MyProfile/MyProfile"
 import { Button, Flex, Text } from "palette"
 import { ImageBackground } from "react-native"
 import { useTracking } from "react-tracking"
