+++ src/app/Scenes/MyProfile/LoggedInUserInfo.tests.tsx
@@ -6,19 +6,19 @@ import { renderWithWrappersLEGACY } from "app/tests/renderWithWrappers"
 import { Text } from "palette"
 import { act } from "react-test-renderer"
 import { createMockEnvironment } from "relay-test-utils"
-import { UserProfileQueryRenderer } from "./LoggedInUserInfo"
+import { MyProfileQueryRenderer } from "./LoggedInUserInfo"
 
 jest.unmock("react-relay")
 const env = defaultEnvironment as any as ReturnType<typeof createMockEnvironment>
 
-describe(UserProfileQueryRenderer, () => {
+describe(MyProfileQueryRenderer, () => {
   it("spins until the operation resolves", () => {
-    const tree = renderWithWrappersLEGACY(<UserProfileQueryRenderer />)
+    const tree = renderWithWrappersLEGACY(<MyProfileQueryRenderer />)
     expect(tree.root.findAllByType(Spinner)).toHaveLength(1)
   })
 
   it("renders upon sucess", () => {
-    const tree = renderWithWrappersLEGACY(<UserProfileQueryRenderer />)
+    const tree = renderWithWrappersLEGACY(<MyProfileQueryRenderer />)
     expect(env.mock.getMostRecentOperation().request.node.operation.name).toBe(
       "LoggedInUserInfoQuery"
     )
@@ -41,7 +41,7 @@ describe(UserProfileQueryRenderer, () => {
   })
 
   it("renders null upon failure", () => {
-    const tree = renderWithWrappersLEGACY(<UserProfileQueryRenderer />)
+    const tree = renderWithWrappersLEGACY(<MyProfileQueryRenderer />)
 
     rejectMostRecentRelayOperation(env, new Error())
